pyproject.toml
[tool.poetry]
name = "najah"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
django-enumfields = "^2.1.1"
factory-boy = "^3.2.1"
django-enumchoicefield = "^3.0.0"
django-extensions = "^3.1.5"
enums = "^0.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
